requirements.txt
Django==5
django-extensions
djangorestframework
django-htmx
django-multiselectfield==0.1.12
#django-auth-ldap==4.0.0
django-modeltranslation==0.18.7
requests
django-environ